Timelock unit tests refactoring
