+++ crates/iota-types/src/unit_tests/timelock/timelock_tests.rs
@@ -31,3 +31,251 @@ fn timelock_ser_deser_roundtrip() {
     );
     assert_eq!(deserialized_timelock.label(), timelock.label());
 }
+
+use std::str::FromStr;
+
+use iota_stardust_sdk::types::block::{
+    address::Ed25519Address,
+    output::{
+        unlock_condition::{
+            AddressUnlockCondition, StorageDepositReturnUnlockCondition, TimelockUnlockCondition,
+        },
+        BasicOutput, BasicOutputBuilder, NativeToken, OutputId, TokenId,
+    },
+};
+
+use crate::timelock::{
+    self,
+    stardust_upgrade_label::STARDUST_UPGRADE_LABEL_VALUE,
+    timelock::{is_timelocked_vested_reward, try_from_stardust, VestedRewardError},
+};
+
+fn vested_reward_output(amount: u64, expiration_time_sec: u32) -> BasicOutput {
+    BasicOutputBuilder::new_with_amount(amount)
+        .add_unlock_condition(AddressUnlockCondition::new(
+            Ed25519Address::from_str(
+                "0xebe40a263480190dcd7939447ee01aefa73d6f3cc33c90ef7bf905abf8728655",
+            )
+            .unwrap(),
+        ))
+        .add_unlock_condition(TimelockUnlockCondition::new(expiration_time_sec).unwrap())
+        .finish()
+        .unwrap()
+}
+
+#[test]
+fn is_timelocked_vested_reward_all_correct() {
+    let output_id = OutputId::from_str(
+        "0xb191c4bc825ac6983789e50545d5ef07a1d293a98ad974fc9498cb18123456780000",
+    )
+    .unwrap();
+    let output = vested_reward_output(10, 1000);
+
+    assert!(is_timelocked_vested_reward(output_id, &output, 100));
+}
+
+#[test]
+fn is_timelocked_vested_reward_min_id() {
+    let output_id = OutputId::from_str(
+        "0xb191c4bc825ac6983789e50545d5ef07a1d293a98ad974fc9498cb18000000000000",
+    )
+    .unwrap();
+    let output = vested_reward_output(10, 1000);
+
+    assert!(is_timelocked_vested_reward(output_id, &output, 100));
+}
+
+#[test]
+fn is_timelocked_vested_reward_max_id() {
+    let output_id = OutputId::from_str(
+        "0xb191c4bc825ac6983789e50545d5ef07a1d293a98ad974fc9498cb18ffffffff0000",
+    )
+    .unwrap();
+    let output = vested_reward_output(10, 1000);
+
+    assert!(is_timelocked_vested_reward(output_id, &output, 100));
+}
+
+#[test]
+fn is_timelocked_vested_reward_incorrect_id() {
+    let output_id = OutputId::from_str(
+        "0xb191c4bc825ac6983789e50545d5ef07a1d293a98ad974fc9498cb17123456780000",
+    )
+    .unwrap();
+    let output = vested_reward_output(10, 1000);
+
+    assert!(!is_timelocked_vested_reward(output_id, &output, 100));
+}
+
+#[test]
+fn is_timelocked_vested_reward_no_timelock_unlock_condition() {
+    let output_id = OutputId::from_str(
+        "0xb191c4bc825ac6983789e50545d5ef07a1d293a98ad974fc9498cb18123456780000",
+    )
+    .unwrap();
+    let output = BasicOutputBuilder::new_with_amount(10)
+        .add_unlock_condition(AddressUnlockCondition::new(
+            Ed25519Address::from_str(
+                "0xebe40a263480190dcd7939447ee01aefa73d6f3cc33c90ef7bf905abf8728655",
+            )
+            .unwrap(),
+        ))
+        .finish()
+        .unwrap();
+
+    assert!(!is_timelocked_vested_reward(output_id, &output, 100));
+}
+
+#[test]
+fn is_timelocked_vested_reward_bigger_milestone_time() {
+    let output_id = OutputId::from_str(
+        "0xb191c4bc825ac6983789e50545d5ef07a1d293a98ad974fc9498cb18123456780000",
+    )
+    .unwrap();
+    let output = vested_reward_output(10, 100);
+
+    assert!(!is_timelocked_vested_reward(output_id, &output, 1000));
+}
+
+#[test]
+fn is_timelocked_vested_reward_same_milestone_time() {
+    let output_id = OutputId::from_str(
+        "0xb191c4bc825ac6983789e50545d5ef07a1d293a98ad974fc9498cb18123456780000",
+    )
+    .unwrap();
+    let output = vested_reward_output(10, 1000);
+
+    assert!(!is_timelocked_vested_reward(output_id, &output, 1000));
+}
+
+#[test]
+fn timelock_from_stardust_all_correct() {
+    let output_id = OutputId::from_str(
+        "0xb191c4bc825ac6983789e50545d5ef07a1d293a98ad974fc9498cb18123456780000",
+    )
+    .unwrap();
+    let output = vested_reward_output(10, 1000);
+
+    let timelock = try_from_stardust(output_id, &output, 100).unwrap();
+
+    assert!(timelock.locked().value() == 10);
+    assert!(timelock.expiration_timestamp_ms() == 1_000_000);
+    assert!(timelock.label().as_ref().unwrap() == STARDUST_UPGRADE_LABEL_VALUE);
+}
+
+#[test]
+fn timelock_from_stardust_with_expired_output() {
+    let output_id = OutputId::from_str(
+        "0xb191c4bc825ac6983789e50545d5ef07a1d293a98ad974fc9498cb18123456780000",
+    )
+    .unwrap();
+    let output = vested_reward_output(10, 1000);
+
+    let err = try_from_stardust(output_id, &output, 1000).unwrap_err();
+
+    assert!(matches!(err, VestedRewardError::UnlockedVestedReward));
+}
+
+#[test]
+fn timelock_from_stardust_with_incorrect_id() {
+    let output_id = OutputId::from_str(
+        "0xb191c4bc825ac6983789e50545d5ef07a1d293a98ad974fc9498cb17123456780000",
+    )
+    .unwrap();
+    let output = vested_reward_output(10, 1000);
+
+    let err = try_from_stardust(output_id, &output, 100).unwrap_err();
+
+    assert!(matches!(err, VestedRewardError::NotVestedReward));
+}
+
+#[test]
+fn timelock_from_stardust_without_timelock_unlock_condition() {
+    let output_id = OutputId::from_str(
+        "0xb191c4bc825ac6983789e50545d5ef07a1d293a98ad974fc9498cb18123456780000",
+    )
+    .unwrap();
+    let output = BasicOutputBuilder::new_with_amount(10)
+        .add_unlock_condition(AddressUnlockCondition::new(
+            Ed25519Address::from_str(
+                "0xebe40a263480190dcd7939447ee01aefa73d6f3cc33c90ef7bf905abf8728655",
+            )
+            .unwrap(),
+        ))
+        .add_unlock_condition(
+            StorageDepositReturnUnlockCondition::new(
+                Ed25519Address::from_str(
+                    "0xebe40a263480190dcd7939447ee01aefa73d6f3cc33c90ef7bf905abf8728655",
+                )
+                .unwrap(),
+                100,
+                100,
+            )
+            .unwrap(),
+        )
+        .finish()
+        .unwrap();
+
+    let err = try_from_stardust(output_id, &output, 1000).unwrap_err();
+
+    assert!(matches!(err, VestedRewardError::NotVestedReward));
+}
+
+#[test]
+fn timelock_from_stardust_extra_unlock_condition() {
+    let output_id = OutputId::from_str(
+        "0xb191c4bc825ac6983789e50545d5ef07a1d293a98ad974fc9498cb18123456780000",
+    )
+    .unwrap();
+    let output = BasicOutputBuilder::new_with_amount(10)
+        .add_unlock_condition(AddressUnlockCondition::new(
+            Ed25519Address::from_str(
+                "0xebe40a263480190dcd7939447ee01aefa73d6f3cc33c90ef7bf905abf8728655",
+            )
+            .unwrap(),
+        ))
+        .add_unlock_condition(TimelockUnlockCondition::new(1000).unwrap())
+        .add_unlock_condition(
+            StorageDepositReturnUnlockCondition::new(
+                Ed25519Address::from_str(
+                    "0xebe40a263480190dcd7939447ee01aefa73d6f3cc33c90ef7bf905abf8728655",
+                )
+                .unwrap(),
+                100,
+                100,
+            )
+            .unwrap(),
+        )
+        .finish()
+        .unwrap();
+
+    let err = try_from_stardust(output_id, &output, 100).unwrap_err();
+
+    assert!(matches!(
+        err,
+        VestedRewardError::UnlockConditionsNumberMismatch
+    ));
+}
+
+#[test]
+fn timelock_from_stardust_with_native_tokens() {
+    let output_id = OutputId::from_str(
+        "0xb191c4bc825ac6983789e50545d5ef07a1d293a98ad974fc9498cb18123456780000",
+    )
+    .unwrap();
+    let output = BasicOutputBuilder::new_with_amount(10)
+        .add_unlock_condition(AddressUnlockCondition::new(
+            Ed25519Address::from_str(
+                "0xebe40a263480190dcd7939447ee01aefa73d6f3cc33c90ef7bf905abf8728655",
+            )
+            .unwrap(),
+        ))
+        .add_unlock_condition(TimelockUnlockCondition::new(1000).unwrap())
+        .add_native_token(NativeToken::new(TokenId::null(), 1).unwrap())
+        .finish()
+        .unwrap();
+
+    let err = try_from_stardust(output_id, &output, 100).unwrap_err();
+
+    assert!(matches!(err, VestedRewardError::NativeTokensNotSupported));
+}
